pyproject.toml
[tool.poetry]
name = "djangorest"
version = "0.1.0"
description = ""
authors = ["hoolywant <sir.saltickow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
coreapi = "^2.3.3"
pyyaml = "^6.0.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
coverage = "^7.3.2"
stripe = "^7.0.0"
env = "^0.1.0"
python-dotenv = "^1.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
eventlet = "^0.33.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
